% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{add}
\alias{add}
\title{Addend SMNG density function}
\usage{
add(j, delta_bar, r, lambda, delta, gamma, beta, mu)
}
\description{
Function that implements the j-th addend of the SMNG density function expressed as an infinite sum.
}
\keyword{internal}
