% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{add_MGF}
\alias{add_MGF}
\title{Addend SMNG moment generating function}
\usage{
add_MGF(j, r = r, lambda, delta, gamma, beta)
}
\description{
Function that implements the j-th addend of the SMNG moment generating function expressed as an infinite sum.
}
\keyword{internal}
