% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPA09}
\alias{EPA09}
\title{Chrysene concentration data}
\format{
Numeric vector.
}
\source{
USEPA. \emph{Statistical analysis of groundwater monitoring data at rcra facilities: Unifed guidance.}
Technical report, Office of Resource Conservation and Recovery, Program Implementation
and Information Division, U.S. Environmental Protection Agency, Washington,
D.C. (2009).
}
\usage{
EPA09
}
\description{
Vector of 8 observations of chrysene concentration (ppb) found in water samples.
}
\keyword{datasets}
