% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatigue}
\alias{fatigue}
\title{Low cycle fatigue data}
\format{
Dataframe with variables:

\code{stress}: stress factor.

\code{cycle}: number of test cycles.
}
\source{
Upadhyay, S. K., and M. Peshwani. \emph{Posterior analysis of lognormal regression models using
the Gibbs sampler.} Statistical Papers 49.1 (2008): 59-85.
}
\usage{
fatigue
}
\description{
Data frame of 22 observations in 2 variables
}
\keyword{datasets}
