% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_quant.R
\name{functional_delta}
\alias{functional_delta}
\title{Target functional to minimize with respect to delta}
\usage{
functional_delta(x, n, quant, gamma, sigma2, lambda, nreg = 1, rel_tol = 1e-04)
}
\description{
Function that evaluate the target functional to minimize allowing
for a different value of the parameter \code{delta}.
}
\keyword{internal}
