% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_quant.R
\name{functional_gamma}
\alias{functional_gamma}
\title{Target functional to minimize with respect to gamma}
\usage{
functional_gamma(x, n, quant, delta, sigma2, lambda, nreg = 1, rel_tol = 1e-04)
}
\description{
Function that evaluate the target functional to minimize
allowing for a different value of the parameter \code{gamma}.
}
\keyword{internal}
