% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_quant.R
\name{g_V}
\alias{g_V}
\title{Integral of the target functional to minimize}
\usage{
g_V(Y, lambda, delta, gamma, quant, sigma2, n, nreg, rel_tol = 1e-04)
}
\description{
Function that implements the integrand of the target functional to minimize in order to
obtain hyperparameters that produces optimal estimates in the frequentist context.
}
\details{
It is implemented exploiting the infinite sum representation of the estimator.
}
\keyword{internal}
