% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_quant.R
\name{g_V_vec}
\alias{g_V_vec}
\title{Vectorization of the function \code{\link{g_V}}}
\usage{
g_V_vec(
  Y,
  lambda,
  delta,
  gamma,
  quant,
  sigma2,
  n,
  nreg = nreg,
  rel_tol = 1e-04
)
}
\description{
It allows to use the integrand of the target functional to
minimize as the argument of the function \code{\link{integrate}}.
}
\keyword{internal}
