% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{inf_sum}
\alias{inf_sum}
\title{Summation SMNG density function}
\usage{
inf_sum(x, lambda, delta, gamma, beta, mu, rel_tol)
}
\description{
Function that adds terms to the infinite sum of the SMNG density function until the relative
error reaches the requested threshold fixed by \code{rel_tol}.
}
\keyword{internal}
