% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{inf_sum_MGF}
\alias{inf_sum_MGF}
\title{Summation SMNG moment generating function}
\usage{
inf_sum_MGF(r, lambda, delta, gamma, beta, rel_tol)
}
\description{
Function that adds terms to the infinite sum of the SMNG moment generating function until the relative
error reaches the requested threshold fixed by \code{rel_tol}.
}
\keyword{internal}
