% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{integral}
\alias{integral}
\title{Integrand SMNG density function}
\usage{
integral(t, a, b, c, l)
}
\description{
Function that implements the integral part that appears in the SMNG density function.
}
\keyword{internal}
