% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{integral_MGF}
\alias{integral_MGF}
\title{Integrand SMNG moment generating function}
\usage{
integral_MGF(t, a, b, c, l)
}
\description{
Function that implements the integral part that appears in the SMNG moment generating function.
}
\keyword{internal}
