% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laminators}
\alias{laminators}
\title{Laminators}
\format{
Dataframe with variables:

\code{Worker}: label of the measured worker.

\code{log_Y}: logarithm of the measured Styrene concentration.
}
\source{
R. H. Lyles, L. L. Kupper, and S. M. Rappaport. \emph{Assessing regulatory compliance of occupational exposures via the balanced one-way random effects ANOVA model} Journal of Agricultural, Biological, and Environmental Statistics (1997).
}
\usage{
laminators
}
\description{
Data frame of 39 observations in 2 variables.
}
\keyword{datasets}
