% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{momentRecursion}
\alias{momentRecursion}
\title{Recursion used for SMNG moments}
\usage{
momentRecursion(center, j, mu, d, g, l, beta)
}
\description{
Recursive forumla that allowas to obtain both the raw and the centered moments starting form the
moments centered with respect to \code{mu}.
}
\keyword{internal}
