% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCBC}
\alias{NCBC}
\title{Naval Construction Battalion Center data}
\format{
Dataframe with 2 variables:

\code{al}: aluminium concentration measures.

\code{mn}: manganese concentration measures.
}
\source{
Singh, Ashok K., Anita Singh, and Max Engelhardt. \emph{The lognormal distribution
in environmental applications.} Technology Support Center Issue Paper. (1997).
}
\usage{
NCBC
}
\description{
Data frame of 17 observations in 2 variables
}
\keyword{datasets}
