% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{RatioBesselK}
\alias{RatioBesselK}
\title{Ratio of Bessel K functions}
\usage{
RatioBesselK(x, nu, nu_diff)
}
\description{
Ratio of Bessel K functions with equal argument. The difference between the two orders is \code{nu_diff}.
}
\keyword{internal}
