% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ReadingTime}
\alias{ReadingTime}
\title{Reading Times data}
\format{
Dataframe with variables:

\code{subj}: label indicating the subject.

\code{item}: label indicating the item read.

\code{so}: variable assuming value 1 (object relative condition) and -1 (subject relative condition).

\code{log_rt}: logarithm of the reading time measured.
}
\source{
E. Gibson and H.-H. I. Wu. \emph{Processing chinese relative clauses in context.} Language and Cognitive Processes, 28(1-2):125-155. (2008).
}
\usage{
ReadingTime
}
\description{
Data frame of 547 observations in 4 variables
}
\keyword{datasets}
