% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{SMNGZmoment}
\alias{SMNGZmoment}
\title{SMNG moments centered in \code{mu}}
\usage{
SMNGZmoment(j, mu, d, g, l, beta)
}
\description{
Function that implements the formula of the SMNG moments ceneterd with respect to the location parameter \code{mu}.
It is used to compute all the moments through the function \code{momentRecursion}
}
\keyword{internal}
