% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_func.R
\name{zerfun}
\alias{zerfun}
\title{Function for finding SMNG quantiles}
\usage{
zerfun(x, mu, delta, gamma, lambda, beta, p, rel_tol)
}
\description{
Non-linear function whose zero can be fuond by the \code{uniroot} procedure in order to obtain the SMNG quantile
corresponding to probability \code{p}.
}
\keyword{internal}
